test(tests): separate integration tests from unit tests
